move lobby code to separate module
